phpmark entry

+++ Php/delete.php
@@ -1,4 +1,5 @@
 <form action="delete.php" method="POST">
+    
 <table border="1">
     <tr>
         <th>Roll no</th>
@@ -21,16 +22,29 @@ if(mysqli_num_rows($b)>0)
         echo "<td>".$row['Name']."</td>";
         echo "<td>".$row['Address']."</td>";
         echo "<td>".$row['Phone']."</td>";
-        echo "<td> <button name='update'>update</button></td>";
+        echo "<td> <button name='update' value='".$row['RollNo']."'>update</button></td>";
         echo "<td> <button name='delete' value='".$row['RollNo']."' >Delete</button></td>";
         echo "</tr>";
+        if(isset($_POST['update'])){
+            $roll=$_POST['update'];
+            header("Location: update.php?roll=$roll");
+            exit;
+
+        }
         if(isset($_POST['delete'])){
         $roll=$_POST['delete'];
         $d = "DELETE FROM studentreg WHERE RollNo=$roll";
-        mysqli_query($con,$d);
+       
+        if( mysqli_query($con,$d)){
+        header("Location: delete.php");
+            exit;
+        } else {
+            echo "error";
+        }
         }
     }
 }
 echo "</table>";
 echo "</form>";
 
+?>
